Satisfy notification Swift file budgets
