+++ .github/swift-file-length-budget.tsv
@@ -2,7 +2,7 @@
 # Format: max_lines<TAB>relative path
 # Reduce counts as files shrink. CI fails if tracked files exceed this budget.
 35469	CLI/cmux.swift
-17854	Sources/AppDelegate.swift
+17850	Sources/AppDelegate.swift
 16274	Sources/ContentView.swift
 15095	Sources/TerminalController.swift
 13161	Sources/Workspace.swift
@@ -32,7 +32,7 @@
 3673	cmuxTests/TabManagerSessionSnapshotTests.swift
 3668	cmuxTests/CLIGenericHookPersistenceTests.swift
 3314	Sources/CmuxConfig.swift
-3124	Sources/Update/UpdateTitlebarAccessory.swift
+2977	Sources/Update/UpdateTitlebarAccessory.swift
 2876	cmuxTests/CMUXOpenCommandTests.swift
 2875	Sources/SessionIndexView.swift
 2581	Sources/KeyboardShortcutSettings.swift
