+++ cmuxUITests/MultiWindowNotificationsUITests.swift
@@ -3,9 +3,9 @@ import Foundation
 import CoreGraphics
 
 final class MultiWindowNotificationsUITests: XCTestCase {
-    private var dataPath = ""
+    var dataPath = ""
     private var socketPath = ""
-    private var launchTag = ""
+    var launchTag = ""
 
     override func setUp() {
         super.setUp()
@@ -153,48 +153,6 @@ final class MultiWindowNotificationsUITests: XCTestCase {
 
         app.typeKey(XCUIKeyboardKey.escape.rawValue, modifierFlags: [])
         XCTAssertTrue(waitForElementToDisappear(targetButton, timeout: 3.0), "Expected popover to close on Escape")
-    }
-
-    func testNotificationsPopoverShowsWorkspaceAsHeadline() {
-        let app = XCUIApplication()
-        app.launchEnvironment["CMUX_UI_TEST_MULTI_WINDOW_NOTIF_SETUP"] = "1"
-        app.launchEnvironment["CMUX_UI_TEST_MULTI_WINDOW_NOTIF_PATH"] = dataPath
-        app.launchEnvironment["CMUX_TAG"] = launchTag
-        launchAllowingHeadlessBackgroundActivation(app)
-        XCTAssertTrue(
-            ensureAppRunningAfterLaunch(app, timeout: 12.0),
-            "Expected app to launch for notification workspace-title test. state=\(app.state.rawValue)"
-        )
-
-        XCTAssertTrue(
-            waitForData(keys: ["notifId1", "workspaceTitle1"], timeout: 15.0),
-            "Expected notification and workspace-title setup data"
-        )
-        guard let setup = loadData(),
-              let notificationId = setup["notifId1"],
-              let workspaceTitle = setup["workspaceTitle1"],
-              !notificationId.isEmpty,
-              !workspaceTitle.isEmpty else {
-            XCTFail("Missing notification workspace-title setup data")
-            return
-        }
-
-        XCTAssertTrue(waitForWindowCount(atLeast: 1, app: app, timeout: 6.0))
-        XCTAssertTrue(
-            ensureAppForegroundForInteraction(app, timeout: 6.0),
-            "Expected cmux to be foreground before opening notifications popover. state=\(app.state.rawValue)"
-        )
-
-        app.typeKey("i", modifierFlags: [.command])
-
-        let workspaceHeadline = app.descendants(matching: .any)
-            .matching(identifier: "NotificationPopoverRow.\(notificationId).workspaceTitle")
-            .firstMatch
-        XCTAssertTrue(
-            workspaceHeadline.waitForExistence(timeout: 6.0),
-            "Expected the notification row to expose its workspace as the headline"
-        )
-        XCTAssertEqual(workspaceHeadline.label, workspaceTitle)
     }
 
     func testNotificationsPopoverJumpToLatestButtonShowsShortcut() {
@@ -696,13 +654,13 @@ final class MultiWindowNotificationsUITests: XCTestCase {
         return waitForFocusChange(from: token, timeout: max(0.0, timeout - firstDeadline))
     }
 
-    private func waitForWindowCount(atLeast count: Int, app: XCUIApplication, timeout: TimeInterval) -> Bool {
+    func waitForWindowCount(atLeast count: Int, app: XCUIApplication, timeout: TimeInterval) -> Bool {
         waitForCondition(timeout: timeout) {
             app.windows.count >= count
         }
     }
 
-    private func launchAllowingHeadlessBackgroundActivation(_ app: XCUIApplication) {
+    func launchAllowingHeadlessBackgroundActivation(_ app: XCUIApplication) {
         let options = XCTExpectedFailure.Options()
         options.isStrict = false
         XCTExpectFailure("App activation may fail on headless CI runners", options: options) {
@@ -710,13 +668,13 @@ final class MultiWindowNotificationsUITests: XCTestCase {
         }
     }
 
-    private func ensureAppRunningAfterLaunch(_ app: XCUIApplication, timeout: TimeInterval) -> Bool {
+    func ensureAppRunningAfterLaunch(_ app: XCUIApplication, timeout: TimeInterval) -> Bool {
         waitForCondition(timeout: timeout) {
             app.state == .runningForeground || app.state == .runningBackground
         }
     }
 
-    private func ensureAppForegroundForInteraction(_ app: XCUIApplication, timeout: TimeInterval) -> Bool {
+    func ensureAppForegroundForInteraction(_ app: XCUIApplication, timeout: TimeInterval) -> Bool {
         if app.state == .runningForeground {
             return true
         }
@@ -747,7 +705,7 @@ final class MultiWindowNotificationsUITests: XCTestCase {
         }
     }
 
-    private func waitForData(keys: [String], timeout: TimeInterval) -> Bool {
+    func waitForData(keys: [String], timeout: TimeInterval) -> Bool {
         waitForCondition(timeout: timeout) {
             guard let data = self.loadData() else { return false }
             return keys.allSatisfy { (data[$0] ?? "").isEmpty == false }
@@ -1533,7 +1491,7 @@ final class MultiWindowNotificationsUITests: XCTestCase {
         return String(data: data, encoding: .utf8)
     }
 
-    private func loadData() -> [String: String]? {
+    func loadData() -> [String: String]? {
         guard let data = try? Data(contentsOf: URL(fileURLWithPath: dataPath)) else {
             return nil
         }
